# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  email: string;
  name: string | null;
  avatarUrl: string | null;
}

export interface AuthMeResponse {
  user: AuthUser | null;
}

export interface AuthLogoutResponse {
  success: boolean;
}

export interface ChatUsageResponse {
  used: number;
  limit: number;
  remaining: number;
  canSend: boolean;
  resetsAt: string;
  isSubscribed: boolean;
}

export interface ChatMessageRequest {
  message: string;
  provider?: string | null;
}

export interface QuotaExceededResponse {
  code: string;
  used: number;
  limit: number;
  resetsAt: string;
}

export interface RazorpayCheckoutResponse {
  orderId: string;
  amount: number;
  currency: string;
  keyId: string;
  userEmail: string;
}

export interface RazorpayVerifyRequest {
  orderId: string;
  paymentId: string;
  signature: string;
}

export interface RazorpayVerifyResponse {
  success: boolean;
}