# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /billing/razorpay/checkout:
    post:
      operationId: createRazorpayCheckout
      tags: [billing]
      summary: Create Razorpay checkout
      responses:
        "200":
          description: Checkout details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RazorpayCheckoutResponse"
        "401":
          description: Not authenticated
  /billing/razorpay/verify:
    post:
      operationId: verifyRazorpayPayment
      tags: [billing]
      summary: Verify Razorpay payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RazorpayVerifyRequest"
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RazorpayVerifyResponse"
        "401":
          description: Not authenticated
components:
  schemas:
...
    RazorpayCheckoutResponse:
      type: object
      properties:
        orderId:
          type: string
        amount:
          type: integer
        currency:
          type: string
        keyId:
          type: string
        userEmail:
          type: string
      required:
        - orderId
        - amount
        - currency
        - keyId
        - userEmail
    RazorpayVerifyRequest:
      type: object
      properties:
        orderId:
          type: string
        paymentId:
          type: string
        signature:
          type: string
      required:
        - orderId
        - paymentId
        - signature
    RazorpayVerifyResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success