# lib/api-zod/src/generated/types/razorpayCheckoutResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RazorpayCheckoutResponse {
  orderId: string;
  amount: number;
  currency: string;
  keyId: string;
  userEmail: string;
}